refactor: follow uicore pattern for webpack exports
- Add barrel file (components/index.js) for importing all components
- Add partial exports for individual component imports
- Move withReduxProvider HOC into component index files
- Remove redundant wrapper files (summit-registration-*.js)
- Keep backward compat index entry pointing to registration-modal
