--- src/summit-registration-form.js
@@ -1,24 +1,0 @@
-/**
- * Copyright 2026 OpenStack Foundation
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * summit-registration-form.js - Standalone registration form entry point
- *
- * This is the entry point for using the registration form on a dedicated page
- * (no modal overlay). For modal usage, use summit-registration-modal.js instead.
- **/
-
-import { withReduxProvider } from './utils/withReduxProvider';
-import RegistrationFormContent from './components/registration-form';
-
-const RegistrationForm = withReduxProvider(RegistrationFormContent);
-
-export default RegistrationForm;
