--- src/summit-registration-modal.js
@@ -1,24 +1,0 @@
-/**
- * Copyright 2026 OpenStack Foundation
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * summit-registration-modal.js - Modal registration widget entry point
- *
- * This is the main entry point for using the registration widget in a modal.
- * For standalone page usage (no modal), use summit-registration-form.js instead.
- **/
-
-import { withReduxProvider } from './utils/withReduxProvider';
-import RegistrationModalContent from './components/registration-modal';
-
-const RegistrationModal = withReduxProvider(RegistrationModalContent);
-
-export default RegistrationModal;
